requirements.txt
# - The OMF runs Ubuntu 18.04 which installs Python 3.6.9 via the "python3-pip" apt package
# - The oldest version of Python available via conda-forge is 3.8.5
# - pip doesn't install python. I must install python before installing with this file
# - timeVAE uses type annotations that are only valid starting with Python 3.9

# - The timeVAE repo is not set up to be pip-installable (i.e. it lacks a setup.py and pyproj.toml)
#package-two @ git+https://github.com/abudesai/timeVAE@main

pandas      # - There's some weird bug when reading H5 files when numpy=2.0. numpy=2.0 is installed when Python=3.9. Therefore, I have to use a newer Python version than 3.9
scipy
setuptools  # - Need setuptools to install timeVAE as a local package
pyyaml
tensorflow
matplotlib
scikit-learn
tables
plotly
# pip install pyarrow       # only needed for parquet support
# pip install fastparquet   # only needed for parquet support